Custom promise classes - assure wait() always return value..
Normally the $unwrap=false parmeter of the
GuzzleHttp\Promise\PromiseInterface::wait($unwrap) should cause null to
be returned by classes implementing the PromiseInterface so the return
type hint should allow null value. Unfortunately when it comes to our
custom promise classes it would break the static code analysis and we
prefer to always return a value.