Search sorting reimplemented.
The only reliable search sorting is to perform it server-side. It anyway
has to be done server side for paging so the question is just if we can
replicate the same order client side. The answer is "clearly no" until
the ordering data isn't explicetly provided as yet another technical
triples (because there's no way to check if e.g. client locale match
server database localle, if their exact collation rules are equal,
etc.).

This commit includes adds generation of ordering triples on the server
side (in the RepoDb class) and sorting according to their values client
side (in the RepoTrait).