RepoDb::getPdoStatementBySqlQuery() tune SearchConfig::$relativesProperties handling
So far only properties listed in SearchConfig::$relativesProperties were
included in the results. Because of that to be able to determine the
relation linking relative with the search match, the relation property
has to be included in the SearchConfig::$relativesProperties list. This
behaviour made it impossible to run the most typical scenario like "give
me only a title of resources pointing to a given one [in a way I can
determine the relation]". Now the linking properties are always
included.