SmartSearch::getFtsQuery() include ILIKE search against the raw column
The full text search based on segments does not work for phrases being a
segment part. This particularly affect searches like parts of an URL
path, a filename extension, etc. Such searches could be easily performed
though using the "raw ILIKE '%phrase%'". The nice part this adjustment
does not change anything in the rest of the smart search workflow
(although it is worth mentioning that such matches are not properly
highlighted by the ts_headline() in the SmartSearch::getSearchPage())

Closes #18